// projects/sigct-ui-ng-lib/src/lib/model/UsagerSanterSocialFichierDTO.ts
export class UsagerSanterSocialFichierDTO {

  id: number;

  idEnregistrement: number;

  idReferenceTypeFichier:  number;

  nom: string;

  description: string;

  titre: string;

  conetnu: string;

  typeContenu: string;

	dateCreation: Date;

	creeUsername: string;

	dateModifie: Date;

  modifieUsername: string;

  file: File;
}